update.sh: download hucon.run to fixed destination path
Previously it ended up at a path based on the current working directory.
When started from the webserver that was `/opt/hucon/webserver/hucon.run`
which got removed by the deploy logic.
So update via web frontend always failed.
Now it is downloaded to `/opt/hucon.run` which is preserved by deploy logic.